displays title screen, most code ported
No it's not at all a Java-style program, but I converted most of the code directly from its C counterpart, with a few changes made to allow Java to display it properly in swing.

Right now it only displays the title logo, but in theory there is support to implement the rest of the game this way.